+++ src/components/RadioButtonGroup/index.js
@@ -1,0 +1,3 @@
+import RadioButtonGroup from './RadioButtonGroup.react';
+
+export default RadioButtonGroup;