+++ src/index.js
@@ -10,6 +10,7 @@ import FlatButton from './components/FlatButton';
 import FontIcon from './components/FontIcon';
 import IconButton from './components/IconButton';
 import Snackbar from './components/Snackbar';
+import RadioButtonGroup from './components/RadioButtonGroup';
 import RaisedButton from './components/RaisedButton';
 import Toggle from './components/Toggle';
 
@@ -26,6 +27,7 @@ export {
   FontIcon,
   IconButton,
   Snackbar,
+  RadioButtonGroup,
   RaisedButton,
   Toggle,
 };